% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{CalcModMatrix}
\alias{CalcModMatrix}
\title{Calculate the Modularity Matrix}
\usage{
CalcModMatrix(input_matrix, is_directed = TRUE, resolution = 1)
}
\arguments{
\item{input_matrix}{A numeric matrix representing the network's adjacency matrix.}

\item{is_directed}{A boolean indicating if the graph is directed (TRUE) or undirected (FALSE).}

\item{resolution}{A numeric value for the resolution parameter (gamma). Default is 1.0.}
}
\value{
A modularity matrix with the same dimensions and names as the input matrix.
}
\description{
This function calculates the modularity matrix from an input adjacency matrix.
It includes validation to ensure the input is a square matrix and that
row and column names match if they exist.
}
