% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{calculate_modularity}
\alias{calculate_modularity}
\title{Function to calculate modularity for a given community structure.}
\usage{
calculate_modularity(
  input_matrix,
  communities,
  is_directed = TRUE,
  resolution = 1
)
}
\arguments{
\item{input_matrix}{A numeric adjacency matrix representing the network.}

\item{communities}{A vector indicating the community membership of each node. The order or names must match the input_matrix.}

\item{is_directed}{A logical value indicating if the network is directed (TRUE) or should be treated as undirected (FALSE).}

\item{resolution}{A numeric value for the resolution parameter (gamma). Default is 1.0.}
}
\value{
A single numeric value representing the modularity score (Q) for the given community partition.
}
\description{
Function to calculate modularity for a given community structure.
}
