% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{flowbca_diagnosis}
\alias{flowbca_diagnosis}
\title{Flow-based Community Analysis Diagnosis}
\usage{
flowbca_diagnosis(flow_input, is_directed = TRUE)
}
\arguments{
\item{flow_input}{The input flow data.}

\item{is_directed}{A logical value indicating whether the graph is directed.}
}
\value{
A list containing the diagnosis statistics.
}
\description{
This function performs a diagnosis of the flow-based community analysis.
}
