% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{flowbca_modularity}
\alias{flowbca_modularity}
\title{Calculate Modularity for each step of flowbca}
\usage{
flowbca_modularity(unit_set, F_matrix_history)
}
\arguments{
\item{unit_set}{A data frame from the result of flowbca function.
It must contain 'sourceunit', 'destinationunit', and 'round' columns.}

\item{F_matrix_history}{A list of flow matrices from the flowbca process.
The first matrix in the list (the original OD matrix) is used for the calculation.}
}
\value{
A data frame with two columns: 'round' and 'modularity',
showing the modularity value for each clustering step.
}
\description{
This function calculates the modularity for each step of the
Flow based-Clustering Algorithm (flowbca) clustering process.
}
\examples{
# Assuming 'bca_result' is an object from a flowbca function call
# modularity_result <- flowbca_modularity(bca_result$unit_set, bca_result$F_matrix_history)
# plot(rev(modularity_result$round$round), modularity_result$round$modularity)
}
