% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{list_to_dendrogram}
\alias{list_to_dendrogram}
\title{Convert a nested list to a dendrogram with labels}
\usage{
list_to_dendrogram(nested_list)
}
\arguments{
\item{nested_list}{Nested list to convert.}
}
\value{
An object of class 'dendrogram'
}
\description{
Converts each hierarchy (name) of a list into branch labels of a dendrogram.
}
