% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{plot_diagnosis}
\alias{plot_diagnosis}
\title{Plot Diagnosis}
\usage{
plot_diagnosis(diagnosis_data, y_var)
}
\arguments{
\item{diagnosis_data}{A list containing relative and absolute data.}

\item{y_var}{Name of the y-axis variable to plot (one of: 'mean', 'min', 'median', 'max', 'intra_flow_ratio', 'inter_flow_ratio', 'g', 'g_rel').}
}
\value{
A plot.
}
\description{
This function plots the diagnosis data.
}
