% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_function.R
\name{remove_holes_sf}
\alias{remove_holes_sf}
\title{Remove Interior Holes from all Geometries in an sf Object}
\usage{
remove_holes_sf(sf_obj)
}
\arguments{
\item{sf_obj}{An \code{sf} object containing POLYGON or MULTIPOLYGON geometries.}
}
\value{
A new \code{sf} object with holes removed from its geometries.
}
\description{
This function iterates over all geometries within an \code{sf} object and removes
any interior holes (islands) from POLYGON and MULTIPOLYGON features. It is a
wrapper around the internal \code{remove_holes} function.
}
